+++ doc/Changelog.md
@@ -3,6 +3,9 @@ Changelog {#Changelog}
 
 # git master {#master}
 
+* [484](https://github.com/Eyescale/Equalizer/pull/484):
+  Implement Window::doneCurrent() to make no OpenGL context current in the
+  current thread
 * [481](https://github.com/Eyescale/Equalizer/pull/481):
   Fix Config::getNextEvent() with definite timeout
 * [467](https://github.com/Eyescale/Equalizer/issues/467):